separate common logic
