+++ rychkov.mihail/common/mem_checker.hpp
